refactor: split `vex-v5-serial` into two crates
